New method, delta calculation
Using the new, more back-prop type method. I have computed the deltas
for each node, in what I think is the correct fashion. Now all that
remains is to use those deltas to produce the change needed at each
weight, for that given instance/class pair. Then, sum these. (Or
different method, if this proves to not make sense)